Move charts to /charts/ folder and install time series adapter
