+++ package.json
@@ -28,6 +28,8 @@
   "dependencies": {
     "@kubernetes/client-node": "^0.20.0",
     "dotenv": "^16.3.1",
+    "chartjs-adapter-date-fns": "^3.0.0",
+    "date-fns": "^2.30.0",
     "express": "^4.18.2",
     "next": "^14.0.3",
     "react": "^18.2.0",
