+++ src/app/dashboard/page.jsx
@@ -1,16 +1,24 @@
-import LinePlot from '../../components/LinePlot'
-import PieChart from '../../components/PieChart'
-import TimeSeriesPlot from '../../components/TimeSeriesPlot'
+import LinePlot from '../charts/LinePlot'
+import PieChart from '../charts/PieChart'
+import TimeSeriesPlot from '../charts/TimeSeriesPlot'
 
 
 
 export default function Page() {
+  const style = { width: "30%" };
+
   return (
   <>
     <h1>Sentry Dashboard</h1>
-    <TimeSeriesPlot />
-    <PieChart />
-    <LinePlot />
+    {/* <div style={style}>
+      <PieChart />
+    </div> */}
+    <div style={style}>
+      <TimeSeriesPlot />
+    </div>
+    {/* <div style={style}>
+      <LinePlot />
+    </div> */}
   </>
   );
 }
